requirements.txt
Brotli==1.0.9
certifi==2022.12.7
flowlauncher==0.2.0
Flox-lib==0.19.2
mutagen==1.46.0
pycryptodomex==3.17
websockets==10.4
yt-dlp==2023.1.6